+++ apps/broker-server/src/main/java/com/google/cloud/broker/apps/brokerserver/accesstokens/providers/DomainWideDelegationAuthorityProvider.java
@@ -30,7 +30,7 @@ public class DomainWideDelegationAuthorityProvider extends AbstractSignedJWTProv
         if (username.length() == 0) {
             throw new IllegalArgumentException();
         }
-        String domain = AppSettings.getInstance().getString(AppSettings.DOMAIN_NAME);
+        String domain = AppSettings.getInstance().getString(AppSettings.GSUITE_DOMAIN);
         return String.format("%s@%s", username, domain);
     }